Add results for Chronos-2-Synth (#89)
This PR add results for the Chronos-2-Synth variant of the Chronos-2
model. As described in the [technical
report](https://arxiv.org/abs/2510.15821) in Sec. 5.4, this variant was
_only_ trained on synthetically generated time series data. The results
can be replicated using the same notebook as Chronos-2 by just changing
the `model_name` to `autogluon/chronos-2-synth` (note: This HF repo is
not public yet. We plan to make it public in the next couple of days).

Also included are minor fixes to the Chronos-2 notebook.